moving npm classes from maven plugin to server module
