+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/common/ThemedURLTranslator.java
@@ -34,6 +34,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.internal.ReflectTools;
+import com.vaadin.flow.server.frontend.ClassPathIntrospector;
 import com.vaadin.flow.theme.AbstractTheme;
 import com.vaadin.flow.theme.Theme;
